Obtain the damond-pancreas-2019 dataset

Obtain the damond-pancreas-2019 dataset, which consists of three data objects: single cell data, multichannel images and cell segmentation masks. The data was obtained by imaging mass cytometry of human pancreas sections from donors with type 1 diabetes.

DamondPancreas2019Data(
  data_type = c("sce", "images", "masks"),
  on_disk = FALSE,
  h5FilesPath = NULL,
  force = FALSE
)

Arguments

data_type
type of data to load, should be `sce` for single cell data, `images` for multichannel images or `masks` for cell segmentation masks.

on_disk
logical indicating if images in form of HDF5Array objects (as .h5 files) should be stored on disk rather than in memory. This setting is valid when downloading images and masks.

h5FilesPath
path to where the .h5 files for on disk representation are stored. This path needs to be defined when on_disk = TRUE. When files should only temporarily be stored on disk, please set h5FilesPath = getHDF5DumpDir()

force
logical indicating if images should be overwritten when files with the same name already exist on disk.

Value

A SingleCellExperiment object with single cell data, a CytoImageList object containing multichannel images, or a CytoImageList object containing cell masks.

Details

This is an Imaging Mass Cytometry (IMC) dataset from Damond et al. (2019), consisting of three data objects:

DamondPancreas2019_images contains a hundred 38-channel images in the form of a CytoImageList class object.

DamondPancreas2019_masks contains the cell segmentation masks associated with the images, in the form of a CytoImageList class object.

DamondPancreas2019_sce contains the single cell data extracted from the images using the cell segmentation masks, as well as the associated metadata, in the form of a SingleCellExperiment. This represents a total of 252,059 cells x 38 channels.

All data are downloaded from ExperimentHub and cached for local re-use. Specific resources can be retrieved by searching for imcdatasets/DamondPancreas2019_sce, imcdatasets/DamondPancreas2019_images, or imcdatasets/DamondPancreas2019_masks.

Mapping between the three data objects is performed via variables located in their metadata columns: mcols() for the CytoImageList objects and ColData() for the SingleCellExperiment object. Mapping at the image level can be performed with the ImageName or ImageNumber variables. Mapping between cell segmentation masks and single cell data is performed with the CellNumber variable, the values of which correspond to the intensity values of the DamondPancreas2019_masks object. For practical examples, please refer to the "Accessing IMC datasets" vignette.

This dataset is a subset of the complete Damond et al. (2019) dataset comprising the data from three pancreas donors at different stages of type 1 diabetes (T1D). The three donors present clearly diverging characteristics in terms of cell type composition and cell-cell interactions, which makes this dataset ideal for benchmarking spatial and neighborhood analysis algorithms.

The assay slot of the SingleCellExperiment object contains two assays:

counts contains mean ion counts per cell.

exprs contains arsinh-transformed counts, with cofactor 1.

The marker-associated metadata, including antibody information and metal tags are stored in the rowData of the SingleCellExperiment object.

The cell-associated metadata are stored in the colData of the SingleCellExperiment object. These metadata include cell types (in colData(sce)$CellType) and broader cell categories, such as "immune" or "islet" cells (in colData(sce)$CellCat). In addition, for cells located inside pancreatic islets, the islet they belong to is indicated in colData(sce)$ParentIslet. For cells not located in islets, the "ParentIslet" value is set to 0 but the spatially closest islet can be identified with colData(sce)$ClosestIslet.

The donor-associated metadata are also stored in the colData of the SingleCellExperiment object. For instance, the donors' IDs can be retrieved with colData(sce)$case and the donors' disease stage can be obtained with colData(sce)$stage.

The three donors present the following characteristics:

6126 is a non-diabetic donor, with large islets containing many beta cells, severe infiltration of the exocrine pancreas with myeloid cells but limited infiltration of islets.

6414 is a donor with recent T1D onset (shortly after diagnosis) showing partial beta cell destruction and mild infiltration of islets with T cells.

6180 is a donor with long-duration T1D (11 years after diagnosis), showing near-total beta cell destruction and limited immune cell infiltration in both the islets and the pancreas.

File sizes:

`DamondPancreas2019_images`: size in memory = 7.4 Gb, size on disk = 1780 Mb.

`DamondPancreas2019_masks`: size in memory = 200.0 Mb, size on disk = 8.6 Mb.

`DamondPancreas2019_sce`: size in memory = 248.6 Mb, size on disk = 145 Mb.

When storing images on disk, these need to be first fully read into memory before writing them to disk. This means the process of downloading the data is slower than directly keeping them in memory. However, downstream analysis will lose its memory overhead when storing images on disk.

Original source: Damond et al. (2019): https://doi.org/10.1016/j.cmet.2018.11.014

Original link to raw data, also containing the entire dataset: https://data.mendeley.com/datasets/cydmwsfztj/2

References

Damond N et al. (2019). A Map of Human Type 1 Diabetes Progression by Imaging Mass Cytometry. Cell Metab 29(3), 755-768.

Author

Nicolas Damond

Examples

sce <- DamondPancreas2019Data(data_type = "sce")

#> Cannot connect to ExperimentHub server, using 'localHub=TRUE' instead

#> snapshotDate(): 2021-05-05

#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation

#> loading from cache

sce

#> class: SingleCellExperiment 
#> dim: 38 252059 
#> metadata(0):
#> assays(2): counts exprs
#> rownames(38): H3 SMA ... Ir191 Ir193
#> rowData names(15): TubeNb MetalTag ... miCAT2 miCAT
#> colnames(252059): E02_1 E02_2 ... J34_1149 J34_1150
#> colData names(26): slide id ... Ethnicity BMI
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):

images <- DamondPancreas2019Data(data_type = "images")

#> Cannot connect to ExperimentHub server, using 'localHub=TRUE' instead

#> snapshotDate(): 2021-05-05

#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation

#> loading from cache

head(images)

#> CytoImageList containing 6 image(s)
#> names(6): E02_a0_full_clean E03_a0_full_clean E04_a0_full_clean E05_a0_full_clean E06_a0_full_clean E07_a0_full_clean 
#> Each image contains 38 channel(s)
#> channelNames(38): H3 SMA INS CD38 CD44 PCSK2 CD99 CD68 MPO SLC2A1 CD20 AMY2A CD3e PPY PIN PD-1 GCG PDX1 SST SYP KRT19 CD45 FOXP3 CD45RA CD8a CA9 IAPP KI-67 NKX6-1 pH3 CD4 CD31 CDH PTPRN pRB cPARP1 Ir191 Ir193 

masks <- DamondPancreas2019Data(data_type = "masks")

#> Cannot connect to ExperimentHub server, using 'localHub=TRUE' instead

#> snapshotDate(): 2021-05-05

#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation

#> loading from cache

head(masks)

#> CytoImageList containing 6 image(s)
#> names(6): E02_a0_full_mask E03_a0_full_mask E04_a0_full_mask E05_a0_full_mask E06_a0_full_mask E07_a0_full_mask 
#> Each image contains 1 channel