Obtain the jackson-fischer-2020 dataset

Obtain the jackson-fischer-2020 dataset, which consists of three data objects: single cell data, multichannel images and cell segmentation masks. The data was obtained by imaging mass cytometry of tumour tissue from patients with breast cancer.

JacksonFischer2020Data(
  data_type = c("sce", "images", "masks"),
  on_disk = FALSE,
  h5FilesPath = NULL,
  force = FALSE
)

Arguments

data_type
type of data to load, should be `sce` for single cell data, `images` for multichannel images or `masks` for cell segmentation masks.

on_disk
logical indicating if images in form of HDF5Array objects (as .h5 files) should be stored on disk rather than in memory. This setting is valid when downloading images and masks.

h5FilesPath
path to where the .h5 files for on disk representation are stored. This path needs to be defined when on_disk = TRUE. When files should only temporarily be stored on disk, please set h5FilesPath = getHDF5DumpDir()

force
logical indicating if images should be overwritten when files with the same name already exist on disk.

Value

A SingleCellExperiment object with single cell data, a CytoImageList object containing multichannel images, or a CytoImageList object containing cell masks.

Details

This is an Imaging Mass Cytometry (IMC) dataset from Jackson, Fischer et al. (2020), consisting of three data objects:

JacksonFischer2020_images contains a hundred 42-channel images in the form of a CytoImageList class object.

JacksonFischer2020_masks contains the cell segmentation masks associated with the images, in the form of a CytoImageList class object.

JacksonFischer2020_sce contains the single cell data extracted from the images using the cell segmentation masks, as well as the associated metadata, in the form of a SingleCellExperiment. This represents a total of 285,851 cells x 42 channels.

All data are downloaded from ExperimentHub and cached for local re-use. Specific resources can be retrieved by searching for imcdatasets/JacksonFischer2020_sce, imcdatasets/JacksonFischer2020_images, or imcdatasets/JacksonFischer2020_masks.

Mapping between the three data objects is performed via variables located in their metadata columns: mcols() for the CytoImageList objects and ColData() for the SingleCellExperiment object. Mapping at the image level can be performed with the ImageNb variable. Mapping between cell segmentation masks and single cell data is performed with the CellNb variable, the values of which correspond to the intensity values of the JacksonFischer2020_masks object. For practical examples, please refer to the "Accessing IMC datasets" vignette.

This dataset is a subset of the complete Jackson, Fischer et al. (2020) dataset comprising the data from tumour tissue from 100 patients with breast cancer (one image per patient).

The assay slot of the SingleCellExperiment object contains three assays:

counts contains mean ion counts per cell.

exprs contains arsinh-transformed counts, with cofactor 1.

quant_norm contains quantile-normalized counts (0 to 1, 99th percentile).

The marker-associated metadata, including antibody information and metal tags are stored in the rowData of the SingleCellExperiment object.

The cell-associated metadata are stored in the colData of the SingleCellExperiment object. These metadata include clusters (in colData(sce)$PhenoGraphBasel) and metaclusters (in colData(sce)$metacluster), as well as spatial information (e.g., cell areas are stored in colData(sce)$Area).

The patient-associated clinical data are also stored in the colData of the SingleCellExperiment object. For instance, the tumor grades can be retrieved with colData(sce)$grade.

File sizes:

`JacksonFischer2020_images`: size in memory = 17.8 Gb, size on disk = 1996 Mb.

`JacksonFischer2020_masks`: size in memory = 433 Mb, size on disk = 10.2 Mb.

`JacksonFischer2020_sce`: size in memory = 517 Mb, size on disk = 272 Mb.

When storing images on disk, these need to be first fully read into memory before writing them to disk. This means the process of downloading the data is slower than directly keeping them in memory. However, downstream analysis will lose its memory overhead when storing images on disk.

Original source: Jackson, Fischer et al. (2020): https://doi.org/10.1038/s41586-019-1876-x

Original link to raw data, containing the entire dataset: https://doi.org/10.5281/zenodo.3518284

References

Jackson, Fischer et al. (2020). The single-cell pathology landscape of breast cancer. Nature 578(7796), 615-620.

Author

Jana Fischer

Examples

sce <- JacksonFischer2020Data(data_type = "sce")

#> snapshotDate(): 2021-05-05

#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation

#> loading from cache

sce

#> class: SingleCellExperiment 
#> dim: 42 285851 
#> metadata(0):
#> assays(3): counts exprs quant_norm
#> rownames(42): EGFR Rutheni_96 ... phospho S6 phospho mTOR
#> rowData names(5): Metal.Tag original metal weight FullStack
#> colnames(285851): BaselTMA_SP41_101_X10Y8_1 BaselTMA_SP41_101_X10Y8_2
#>   ... BaselTMA_SP43_97_X16Y6_3647 BaselTMA_SP43_97_X16Y6_3648
#> colData names(74): core id ... response ImageNb
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):

images <- JacksonFischer2020Data(data_type = "images")

#> snapshotDate(): 2021-05-05

#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation

#> loading from cache

head(images)

#> CytoImageList containing 6 image(s)
#> names(6): BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_106_18_X13Y5_248_a0_full BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_107_114_X13Y4_219_a0_full BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_108_117_X13Y3_187_a0_full BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_114_42_X14Y5_250_a0_full BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_122_166_X15Y4_231_a0_full BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_123_86_X15Y3_203_a0_full 
#> Each image contains 42 channel(s)
#> channelNames(42): Rutheni_96 Rutheni_98 Rutheni_99 Rutheni_100 Rutheni_101 Rutheni_102 Rutheni_104 Histone H3K27me3 CK5 Fibrone CK19 CK8/18 Twist CD68 CK14 SMA Vimenti cMyc HER2 CD3epsi phospho Histone Slug Estroge Progest p53 CD44 CD45 GATA3 CD20 CAIX ECadhe Ki67 EGFR phospho S6 vWF phospho mTOR CK7 panCyto CC3/cPARP Iridium_191 Iridium_193 

masks <- JacksonFischer2020Data(data_type = "masks")

#> snapshotDate(): 2021-05-05

#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation

#> loading from cache

head(masks)

#> CytoImageList containing 6 image(s)
#> names(6): BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_106_18_X13Y5_248_a0_full_maks BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_107_114_X13Y4_219_a0_full_maks BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_108_117_X13Y3_187_a0_full_maks BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_114_42_X14Y5_250_a0_full_maks BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_122_166_X15Y4_231_a0_full_maks BaselTMA_SP41_15.475kx12.665ky_10000x8500_5_20170905_123_86_X15Y3_203_a0_full_maks 
#> Each image contains 1 channel