Obtain the zanotelli-spheroids-2020 dataset

Obtain the zanotelli-spheroids-2020 dataset, which consists of three data objects: single cell data, multichannel images and cell segmentation masks. The data were obtained by imaging mass cytometry of sections of 3D spheroids generated from different cell lines.

ZanotelliSpheroids2020Data(
  data_type = c("sce", "images", "masks"),
  on_disk = FALSE,
  h5FilesPath = NULL,
  force = FALSE
)

Arguments

data_type
type of data to load, should be sce for single cell data, images for multichannel images or masks for cell segmentation masks.

on_disk
logical indicating if images in form of HDF5Array objects (as .h5 files) should be stored on disk rather than in memory. This setting is valid when downloading images and masks.

h5FilesPath
path to where the .h5 files for on disk representation are stored. This path needs to be defined when on_disk = TRUE. When files should only temporarily be stored on disk, please set h5FilesPath = getHDF5DumpDir()

force
logical indicating if images should be overwritten when files with the same name already exist on disk.

Value

A SingleCellExperiment object with single cell data, a CytoImageList object containing multichannel images, or a CytoImageList object containing cell segmentation masks.

Details

This is an Imaging Mass Cytometry (IMC) dataset from Zanotelli et al. (2020), consisting of three data objects:

ZanotelliSpheroids2020_images contains 517 multichannel images, each containing 51 channels, in the form of a CytoImageList class object.

ZanotelliSpheroids2020_masks contains the cell segmentation masks associated with the images, in the form of a CytoImageList class object.

ZanotelliSpheroids2020_sce contains the single cell data extracted from the images using the cell segmentation masks, as well as the associated metadata, in the form of a SingleCellExperiment. This represents a total of 229,047 cells x 51 channels.

All data are downloaded from ExperimentHub and cached for local re-use. Specific resources can be retrieved by searching for imcdatasets/ZanotelliSpheroids2020_sce, imcdatasets/ZanotelliSpheroids2020_images, or imcdatasets/ZanotelliSpheroids2020_masks.

Mapping between the three data objects is performed via variables located in their metadata columns: mcols() for the CytoImageList objects and ColData() for the SingleCellExperiment object. Mapping at the image level can be performed with the ImageName or ImageNumber variables. Mapping between cell segmentation masks and single cell data is performed with the CellNumber variable, the values of which correspond to the intensity values of the ZanotelliSpheroids2020_masks object. For practical examples, please refer to the "Accessing IMC datasets" vignette.

This dataset was obtained as following (the names of the experimental variables, located in the colData of the SingleCellExperiment object, are indicated in parentheses): i) Cells from four different cell lines (cellline) were seeded at three different densities (concentration, relative densities) and grown for either 72 or 96 hours (time_point, duration in hours). In the appropriate experimental conditions (see the paper for details), the cells aggregate into 3D spheroids. ii) Cells were harvested and pooled into 60-well barcoding plates. iii) A pellet of each spheroid pool was generated and cut into several 6 um-thick sections. iv) A subset of these sections (site_id) were stained with an IMC panel and acquired as one or more acquisitions (acquisition_id) containing multiple spheres each. v) Spheres in these acquisitions were identified by computer vision and cropped into individual images (ImageNumber).

Other relevant cell metadata include:

condition_name: experimental conditions in the format: "Cell line name"_c"seeding density"_tp"time point".

Center_X/Y: object centroid position in image.

Area: area of the cell (um^2).

dist.rim: estimated distance to spheroid border.

dist.sphere: distance to spheroid section border.

dist.other: distance to the closest of the other spheroid sections in the same image (if there is any).

dist.bg: distance to background pixels.

counts_neighb: contains arsinh-transformed counts, with cofactor 1.

exprs_neighb: contains arsinh-transformed counts, with cofactor 1.

For a full description of the other experimental variables, please refer to the publication (https://doi.org/10.15252/msb.20209798) and to the original dataset repository (https://doi.org/10.5281/zenodo.4271910).

The marker-associated metadata, including antibody information and metal tags are stored in the rowData of the SingleCellExperiment object. The channels with names starting with "BC_" are the channels used for barcoding. Post-transcriptional modification of the protein targets are indicated in brackets.

The assay slot of the SingleCellExperiment object contains four assays:

counts: mean ion counts per cell.

exprs: arsinh-transformed counts per cell, with cofactor 1.

counts_neighb: mean ion counts of the neighboring cells.

exprs_neighb: arsinh-transformed counts (cofactor 1) of the neighboring cells.

The metadata slot of the SingleCellExperiment object contains a graph of cell neighbors, generated with the igraph::graph_from_data_frame function.

File sizes:

`ZanotelliSpheroids2020_images`: size in memory = 21.2 Gb, size on disk = 881 Mb.

`ZanotelliSpheroids2020_masks`: size in memory = 426 Mb, size on disk = 11.6 Mb.

`ZanotelliSpheroids2020_sce`: size in memory = 584 Mb, size on disk = 340 Mb.

When storing images on disk, these need to be first fully read into memory before writing them to disk. This means the process of downloading the data is slower than directly keeping them in memory. However, downstream analysis will lose its memory overhead when storing images on disk.

Original source: Zanotelli et al. (2020): https://doi.org/10.15252/msb.20209798

Original link to raw data, also containing the entire dataset: https://doi.org/10.5281/zenodo.4271910

References

Zanotelli VRT et al. (2020). A quantitative analysis of the interplay of environment, neighborhood, and cell state in 3D spheroids Mol Syst Biol 16(12), e9798.

Author

Nicolas Damond

Examples

sce <- ZanotelliSpheroids2020Data(data_type = "sce")

#> snapshotDate(): 2021-05-05

#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation

#> loading from cache

sce

#> class: SingleCellExperiment 
#> dim: 51 229047 
#> metadata(1): graph
#> assays(4): counts counts_neighb exprs exprs_neighb
#> rownames(51): p-p70S6K [T389] p-Stat3 [Y705] ... p-AKT [S473] BC_Pd105
#> rowData names(6): ref_plane_number metal ... Antibody.Clone is_cc
#> colnames(229047): 454 455 ... 349952 349968
#> colData names(30): ImageNumber Row.names ... mask_filename_cell
#>   ImageName
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):

images <- ZanotelliSpheroids2020Data(data_type = "images")

#> snapshotDate(): 2021-05-05

#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation

#> loading from cache

head(images)

#> CytoImageList containing 6 image(s)
#> names(6): 20190619_p173_slide31_ac1_vz_s1_p1_r11_a11_ac_full_l2_x217_y72_comp 20190619_p173_slide31_ac1_vz_s1_p1_r13_a13_ac_full_l1_x0_y0_comp 20190619_p173_slide31_ac1_vz_s1_p1_r14_a14_ac_full_l1_x0_y127_comp 20190619_p173_slide31_ac1_vz_s1_p1_r14_a14_ac_full_l2_x277_y0_comp 20190619_p173_slide31_ac1_vz_s1_p1_r14_a14_ac_full_l3_x389_y144_comp 20190619_p173_slide31_ac1_vz_s1_p1_r15_a15_ac_full_l1_x0_y0_comp 
#> Each image contains 51 channel(s)
#> channelNames(51): BC_Pd102 BC_Rh103 BC_Pd104 BC_Pd105 BC_Pd106 BC_Pd108 BC_Pd110 BC_In113 BC_In115 Telox2 meHH3 [K27] HH3 EGFR p-FAK [Y397] p-MEK1/2 [S217/S221] p-MAPKAPK2 [T334] p-p70S6K [T389] p-AuroraB [T232] p-Stat1 [S727] p-p53 [S15] CD44 p-EGFR [Tyr1173] p-AMPKalpha [T172] p-HH3 [S28] p-Erk1/2 [T202/Y204] p-Her2 [Y1196] p-p38 [T180/Y182] p-Gsk3 [S9] HER2 Survivin CyclinB1 Vimentin p-AKT [S473] AKT p-CDK1 [Y15] CarbonicAnhydraseIX GFP_Er167 Ki-67 GFP_Tm169 p-JNK [T183/Y185] p-S6 [S235/S236] cleaved-Caspase3 p-Stat3 [Y705] cleaved-PARP p-Rb [S807/S811] DYKDDDDK DNA BC_Pt194 BC_Pt195 BC_Bi209 BC_Y89 

masks <- ZanotelliSpheroids2020Data(data_type = "masks")

#> Cannot connect to ExperimentHub server, using 'localHub=TRUE' instead

#> snapshotDate(): 2021-05-05

#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation

#> loading from cache

head(masks)

#> CytoImageList containing 6 image(s)
#> names(6): 20190619_p173_slide31_ac1_vz_s1_p1_r11_a11_ac_full_l2_x217_y72_cell 20190619_p173_slide31_ac1_vz_s1_p1_r13_a13_ac_full_l1_x0_y0_cell 20190619_p173_slide31_ac1_vz_s1_p1_r14_a14_ac_full_l1_x0_y127_cell 20190619_p173_slide31_ac1_vz_s1_p1_r14_a14_ac_full_l2_x277_y0_cell 20190619_p173_slide31_ac1_vz_s1_p1_r14_a14_ac_full_l3_x389_y144_cell 20190619_p173_slide31_ac1_vz_s1_p1_r15_a15_ac_full_l1_x0_y0_cell 
#> Each image contains 1 channel